+++ src/main/java/insane96mcp/insanelib/module/base/betterfallingblocks/BetterFallingBlockExtensor.java
@@ -2,7 +2,7 @@ package insane96mcp.insanelib.module.base.betterfallingblocks;
 
 import net.minecraft.world.entity.Entity;
 
-public interface BetterFallingBlockAccessor {
+public interface BetterFallingBlockExtensor {
     void insanelib$setSource(Entity source);
     Entity insanelib$getSource();
 }